+++ text/atlas_test.go
@@ -3,7 +3,8 @@ package text_test
 import (
 	"testing"
 
-	"github.com/faiface/pixel/text"
+	"github.com/duysqubix/pixel2/text"
+	"golang.org/x/image/font/inconsolata"
 )
 
 func TestAtlas7x13(t *testing.T) {
@@ -21,4 +22,8 @@ func TestAtlas7x13(t *testing.T) {
 			}
 		}
 	}
+}
+
+func TestAtlasInconsolata(t *testing.T) {
+	text.NewAtlas(inconsolata.Regular8x16, text.ASCII)
 }