+++ text/text.go
@@ -6,7 +6,7 @@ import (
 	"unicode"
 	"unicode/utf8"
 
-	"github.com/faiface/pixel"
+	"github.com/duysqubix/pixel2"
 	"golang.org/x/image/font/basicfont"
 )
 
@@ -101,6 +101,7 @@ type Text struct {
 	trans  pixel.TrianglesData
 	transD pixel.Drawer
 	dirty  bool
+	anchor pixel.Anchor
 }
 
 // New creates a new Text capable of drawing runes contained in the provided Atlas. Orig and Dot
@@ -135,6 +136,7 @@ func New(orig pixel.Vec, atlas *Atlas) *Text {
 
 	txt.transD.Picture = txt.atlas.pic
 	txt.transD.Triangles = &txt.trans
+	txt.transD.Cached = true
 
 	txt.Clear()
 
@@ -180,6 +182,12 @@ func (txt *Text) BoundsOf(s string) pixel.Rect {
 	}
 
 	return bounds
+}
+
+// AlignedTo returns the text moved by the given anchor.
+func (txt *Text) AlignedTo(anchor pixel.Anchor) *Text {
+	txt.anchor = anchor
+	return txt
 }
 
 // Clear removes all written text from the Text. The Dot field is reset to Orig.
@@ -244,6 +252,10 @@ func (txt *Text) DrawColorMask(t pixel.Target, matrix pixel.Matrix, mask color.C
 		txt.mat = matrix
 		txt.dirty = true
 	}
+
+	offset := txt.Orig.Sub(txt.Bounds().Max.Add(txt.Bounds().AnchorPos(txt.anchor.Opposite())))
+	txt.mat = pixel.IM.Moved(offset).Chained(txt.mat)
+
 	if mask == nil {
 		mask = pixel.Alpha(1)
 	}