+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Multiple Window Management Framework
+
+## [v0.12.0] 2023-10-02
+- Add AnchorPos struct and functions #252
+- Add Clipboard Support
+- Fix SIGSEGV on text.NewAtlas if glyph absent 
+- Use slice for range in Drawer.Dirty(), to improve performance
+- GLTriangle's fragment shader is used when rendered by the Canvas.
+- Add MSAA support
+
+## [v0.10.0] 2020-08-22
+- Add AnchorPos struct and functions
 - Gamepad API added
 - Support setting an initial window position
 - Support hiding the window initially
@@ -46,7 +58,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Changelog for this and older versions can be found on the corresponding [GitHub
 releases](https://github.com/faiface/pixel/releases).
 
-[Unreleased]: https://github.com/faiface/pixel/compare/v0.10.0-beta...HEAD
+[Unreleased]: https://github.com/duysqubix/pixel2/compare/v0.12.0...HEAD
+[v0.12.0]: https://github.com/faiface/pixel/compare/v0.12.0...dev
+[v0.10.0]: https://github.com/faiface/pixel/compare/v0.10.0-beta...v0.10.0
 [v0.10.0-beta]: https://github.com/faiface/pixel/compare/v0.10.0-alpha...v0.10.0-beta
 [v0.10.0-alpha]: https://github.com/faiface/pixel/compare/v0.9.0...v0.10.0-alpha
 [v0.9.0]: https://github.com/faiface/pixel/compare/v0.8.0...v0.9.0